% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential-expression.R
\name{addDifferentialExpression}
\alias{addDifferentialExpression}
\title{Manually add differential expression results to a MirnaExperiment object}
\usage{
addDifferentialExpression(
  mirnaObj,
  mirnaDE,
  geneDE,
  significantMirnas,
  significantGenes
)
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{mirnaDE}{A \code{data.frame} containing the output of miRNA differential
expression analysis. Check the \emph{details} section to see the required format}

\item{geneDE}{A \code{data.frame} containing the output of gene differential
expression analysis. Check the \emph{details} section to see the required format}

\item{significantMirnas}{A \code{character} vector containing the IDs of
statistically differentially expressed miRNAs. See the \emph{details} section for
further information}

\item{significantGenes}{A \code{character} vector containing the IDs of
statistically differentially expressed genes. See the \emph{details} section for
further information}
}
\value{
A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object containing differential
expression results. To access these results, the user may run the
\code{\link[=mirnaDE]{mirnaDE()}} and \code{\link[=geneDE]{geneDE()}} functions for miRNAs and genes, respectively.
}
\description{
This function allows to add miRNA and gene differential expression results
to a \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object. Instead of running
\code{\link[=performMirnaDE]{performMirnaDE()}} and \code{\link[=performGeneDE]{performGeneDE()}} functions, this one allows to use
differential expression analyses carried out in other ways. This is
particularly useful in order to use the pipeline implemented in MIRit for
proteomic data and for expression data deriving from different technologies.
}
\details{
The following paragraphs briefly explain the formats needed for mirnaDE,
geneDE, significantMirnas and significantGenes.
\subsection{mirnaDE and geneDE}{

\code{mirnaDE} and \code{geneDE} are two objects of class \code{data.frame} containing
the results of miRNA and gene differential expression analysis respectively.
These tables should contain the differential expression results for all
miRNAs/genes analyzed, not just for statistically significant species.
All \code{data.frame} objects can be used, as long as they have:
\itemize{
\item One column containing miRNA/gene names (according to miRBase/hgnc
nomenclature). Accepted column names are: \code{ID}, \code{Symbol}, \code{Gene_Symbol},
\code{Mirna}, \code{mir}, \code{Gene}, \code{gene.symbol}, \code{Gene.symbol};
\item One column with log2 fold changes. Accepted column names are: \code{logFC},
\code{log2FoldChange}, \code{FC}, \code{lFC};
\item One column with average expression. Accepted column names are: \code{AveExpr},
\code{baseMean}, \code{logCPM};
\item One column with the p-values resulting from the differential expression
analysis. Accepted column names are: \code{P.Value}, \code{pvalue}, \code{PValue},
\code{Pvalue};
\item One column containing p-values adjusted for multiple testing. Accepted
column names are: \code{adj.P.Val}, \code{padj}, \code{FDR}, \code{fdr}, \code{adj}, \code{adj.p}, \code{adjp}.
}
}

\subsection{significantMirnas and significantGenes}{

\code{significantMirnas} and \code{significantGenes} are two \code{character} vectors that
specifies the IDs of miRNAs and genes considered to be significantly
differentially expressed. The miRNA IDs contained in \code{significantMirnas}
must be present in the \code{ID} column of \code{mirnaDE}, in the same way as gene
symbols in \code{significantGenes} must be present in the \code{ID} column of \code{geneDE}.
}
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# create samples metadata
meta <- data.frame("primary" = colnames(geneCounts),
"mirnaCol" = colnames(mirnaCounts), "geneCol" = colnames(geneCounts),
"disease" = c(rep("PTC", 8), rep("NTH", 8)), 
"patient" = c(rep(paste("Sample_", seq(8), sep = ""), 2)))

# perform miRNA DE with DESeq2 separately
dds_m <- DESeq2::DESeqDataSetFromMatrix(countData = obj[["microRNA"]],
colData = meta, design = ~ 0 + disease + patient)
dds_m <- DESeq2::DESeq(dds_m)
de_m <- as.data.frame(DESeq2::results(dds_m,
contrast = c("disease", "PTC", "NTH"), pAdjustMethod = "fdr"))

# perform gene DE with DESeq2 separately
dds_g <- DESeq2::DESeqDataSetFromMatrix(countData = obj[["genes"]],
colData = meta, design = ~ 0 + disease + patient)
dds_g <- DESeq2::DESeq(dds_g)
de_g <- as.data.frame(DESeq2::results(dds_g,
contrast = c("disease", "PTC", "NTH"), pAdjustMethod = "fdr"))

# prepare DE tables
de_m$ID <- rownames(de_m)
de_m <- na.omit(de_m)
de_g$ID <- rownames(de_g)
de_g <- na.omit(de_g)

# define significant features
sig_m <- de_m$ID[de_m$padj < 0.05, ]
sig_g <- de_g$ID[de_g$padj < 0.05, ]

# add DE results to MirnaExperiment object
obj <- addDifferentialExpression(obj, de_m, de_g, sig_m, sig_g)

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
