% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{enrichmentBarplot}
\alias{enrichmentBarplot}
\title{Create a barplot for functional enrichment analysis}
\usage{
enrichmentBarplot(
  enrichment,
  showTerms = 10,
  splitDir = TRUE,
  ordBy = "ratio",
  colBy = "padj",
  title = NULL
)
}
\arguments{
\item{enrichment}{An object of class
\code{\link[=FunctionalEnrichment-class]{FunctionalEnrichment}} containing
enrichment results}

\item{showTerms}{It is the number of top significant terms to show or,
alternatively, a character vector indicating the enriched terms to plot.
Default is \code{10}}

\item{splitDir}{Logical, if \code{TRUE} the resulting plot will be divided in
two columns on the basis of enrichment direction (Up and Down).
Default is \code{TRUE}. This only applies if enrichment method is GSEA or CAMERA}

\item{ordBy}{The parameter used to set the x-axis scale. It must be one of
\code{ratio} (default), \code{padj}, \code{pval} and \code{overlap}}

\item{colBy}{The parameter used to set the color scale. It must be one of
\code{ratio}, \code{padj} (default), \code{pval} and \code{overlap}}

\item{title}{The title of the plot. Default is \code{NULL} not to include a plot
title}
}
\value{
A \code{ggplot} graph with a barplot of enrichment results.
}
\description{
This function produces a barplot to show the results of functional
enrichment analyses carried out through over-representation analysis (ORA),
gene set enrichment analysis (GSEA), and competitive gene set test accounting
for inter-gene correlation (CAMERA). In particular, this function can take
as input enrichment results generated by \code{\link[=enrichGenes]{enrichGenes()}} and
\code{\link[=enrichMirnas]{enrichMirnas()}} functions.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# perform GSEA with KEGG database
gse <- enrichGenes(obj, method = "GSEA",
database = "KEGG", organism = "Homo sapiens")

# extract results
res <- enrichmentResults(gse)

# plot results
enrichmentBarplot(gse)

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
