% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllClasses.R
\name{enrichmentResults}
\alias{enrichmentResults}
\alias{enrichmentResults,enrichResult-method}
\alias{enrichmentResults,gseaResult-method}
\alias{enrichmentResults,MirnaEnrichment-method}
\alias{enrichmentResults,MirnaGsea-method}
\title{Extract results from enrichment objects}
\usage{
enrichmentResults(object)

\S4method{enrichmentResults}{enrichResult}(object)

\S4method{enrichmentResults}{gseaResult}(object)

\S4method{enrichmentResults}{MirnaEnrichment}(object)

\S4method{enrichmentResults}{MirnaGsea}(object)
}
\arguments{
\item{object}{An object of class \code{\link[=MirnaEnrichment-class]{MirnaEnrichment}}
or \code{\link[=MirnaGsea-class]{MirnaGsea}} containing miRNA enrichment results.
Alternatively, object of class
\code{\link[DOSE:enrichResult-class]{enrichResult-class}} or
\code{\link[DOSE:gseaResult-class]{gseaResult-class}} are also accepted}
}
\value{
A \code{data.frame} object with the full results of the enrichment analysis.
}
\description{
This function allows to access the \code{result} slot of
\code{\link[=MirnaEnrichment-class]{MirnaEnrichment}} and \code{\link[=MirnaGsea-class]{MirnaGsea}}
objects. This is useful to take a closer look at all the
enriched terms of an enrichment analysis. Note that this function can also
extract the enrichment results for objects of classes
\code{\link[DOSE:enrichResult-class]{enrichResult-class}} and
\code{\link[DOSE:gseaResult-class]{gseaResult-class}}.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# perform a GSEA with the miEAA category GO Annotations
gse <- gseaMirnas(obj, organism = "Homo sapiens",
category = "GO_Annotations_mature")

# extract results
res <- enrichmentResults(gse)

# plot results
mirnaDotplot(gse)

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
