% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichMirnas}
\alias{enrichMirnas}
\title{Perform an over-representation analysis (ORA) for DE-miRNAs}
\usage{
enrichMirnas(
  mirnaObj,
  organism = "Homo sapiens",
  category = "all",
  pCutoff = 0.05,
  pAdjustment = "fdr",
  independentAdj = TRUE,
  minHits = 2,
  mirnaType = "mature",
  sortBy = "category",
  convertVer = FALSE,
  mirBaseVer = 22,
  ...
)
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{organism}{The name of the organism under consideration. It must be one
of: \verb{Homo sapiens}, \verb{Mus musculus}, \verb{Rattus norvegicus},
\verb{Arabidopsis thaliana}, \verb{Bos taurus}, \verb{Caenorhabditis elegans},
\verb{Drosophila melanogaster}, \verb{Danio rerio}, \verb{Gallus gallus}, \verb{Sus scrofa}}

\item{category}{The name of the miEAA 2.0 category to use for the enrichment
analysis. The list of available categories can be obtained through the
\code{\link[=listCategories]{listCategories()}} function. Moreover, the user can also perform the
enrichment analysis on all miEAA 2.0 categories by specifying \code{all} (default)}

\item{pCutoff}{The adjusted p-value cutoff to use for statistical
significance. The default value is \code{0.05}}

\item{pAdjustment}{The p-value correction method for multiple testing. It
must be one of: \code{fdr} (default), \code{none}, \code{holm}, \code{hochberg}, \code{hommel},
\code{bonferroni}, \code{BY}}

\item{independentAdj}{Logical, whether to adjust the p-values of each
category independently or to adjust p-values collectively over all
categories. Default is \code{TRUE}}

\item{minHits}{The minimum number of miRNAs that must be present in a
subcategory in order to include that subcategory in the results
(default is \code{2})}

\item{mirnaType}{The type of miRNA specie: use \code{mature} for the enrichment
of mature microRNAs; use \code{precursor}, for the enrichment analysis of miRNA
precursors}

\item{sortBy}{The column used to sort the results table. It must be one of:
\code{category} (default), \code{subcategory}, \code{enrichment}, \code{p_value}, \code{p_adjusted},
\code{q_value}, \code{observed}}

\item{convertVer}{Logical, whether miRNA IDs stored in \code{mirnaObj} need to
be converted to miRBase version 22. Default is \code{FALSE}}

\item{mirBaseVer}{The miRBase version of miRNA IDs stored in \code{mirnaObj}.
This parameter is only needed if miRNA IDs need to be converted to
version 22}

\item{...}{Other arguments that can be passed to
\code{\link[rbioapi:rba_mieaa_enrich]{rbioapi::rba_mieaa_enrich()}}}
}
\value{
A \code{list} object containing miEAA 2.0 enrichment results. This \code{list} object
has two \code{\link[=MirnaEnrichment-class]{MirnaEnrichment}} objects, one for the
enrichment results of up-regulated miRNAs and one for enrichment results of
down-regulated miRNAs. To access one or the other, the user can make use of
\code{res[["upregulated"]]} and \code{res[["downregulated"]]}. Further, it is possible
to access the results of these \code{\link[=MirnaEnrichment-class]{MirnaEnrichment}}
objects through the \code{\link[=enrichmentResults]{enrichmentResults()}} function. For additional
instructions see \code{\link[=MirnaEnrichment-class]{MirnaEnrichment}} class.
}
\description{
This function uses miEAA 2.0 (Kern et al., 2020) to perform an
over-representation analysis of differentially expressed miRNAs in order to
investigate the biological consequences of miRNA dysregulations. This can be
pivotal for understanding the biological pathways affected by
miRNA alterations.
}
\details{
To understand the functions of DE-miRNAs, one of the most common approaches
is to retrieve DE-miRNA targets and perform the enrichment of these genes.
However, Bleazard et al. discovered that this approach may lead to biased
results, thus limiting the reliability of the enrichment analysis.
To overcome this limit, several other resources were developed, including
miEAA 2.0, which is a tool based on GeneTrail that facilitates the
interpretation of microRNA dysregulations.
This tool allows to perform enrichment analyses with multiple categories for
assessing different biological features.
These categories, which may not be available for some species,
include databases like miRWalk, mirPathDB and MNDR, and allow to investigate
different features such as biological processes, molecular functions,
biological pathways, diseases, target genes, organs, tissues and more. To
retrieve the complete list of available miEAA 2.0 categories, use the
\code{\link[=listCategories]{listCategories()}} function.
}
\note{
To query miEAA 2.0, this function uses the package \code{rbioapi}, which
provides convenient methods for accessing miEAA 2.0 to perform
enrichment analyses.
}
\examples{
# perform an over-represantation analysis with the miEAA category miRWalk
enr <- enrichMirnas(obj, organism = "Homo sapiens",
category = "miRWalk_GO_mature")

# divide the enrichment of up- and down-regulated miRNAs
enr_up <- enr[["upregulated"]]
enr_down <- enr[["downregulated"]]

# extract results
res_up <- enrichmentResults(enr_up)
res_down <- enrichmentResults(enr_down)

# plot results
mirnaDotplot(enr_up)
mirnaDotplot(enr_down)

}
\references{
Fabian Kern, Tobias Fehlmann, Jeffrey Solomon, Louisa Schwed, Nadja Grammes,
Christina Backes, Kendall Van Keuren-Jensen, David Wesley Craig, Eckart
Meese, Andreas Keller, miEAA 2.0: integrating multi-species microRNA
enrichment analysis and workflow management systems, Nucleic Acids Research,
Volume 48, Issue W1, 02 July 2020, Pages W521–W528,
\url{https://doi.org/10.1093/nar/gkaa309}

Thomas Bleazard, Janine A Lamb, Sam Griffiths-Jones, Bias in microRNA
functional enrichment analysis, Bioinformatics, Volume 31, Issue 10, 15 May
2015, Pages 1592–1598, \url{https://doi.org/10.1093/bioinformatics/btv023}

Moosa Rezwani, Ali Akbar Pourfathollah, Farshid Noorbakhsh, rbioapi:
user-friendly R interface to biologic web services’ API, Bioinformatics,
Volume 38, Issue 10, 15 May 2022, Pages 2952–2953,
\url{https://doi.org/10.1093/bioinformatics/btac172}
}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
