% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{findMirnaSNPs}
\alias{findMirnaSNPs}
\title{Find disease-associated SNPs occurring at DE-miRNA loci}
\usage{
findMirnaSNPs(mirnaObj, diseaseUML, database = "ALL", score = c(0, 1))
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{diseaseUML}{The CUI identifier of a disease of interest. This can be
identified with the \code{\link[=searchDisease]{searchDisease()}} function}

\item{database}{The name of the database of disease-SNPs associations to
draw from. The default value is \code{ALL} to use all databases included in
DisGeNet. For a detailed list of all the possible values, see the details
section}

\item{score}{A numeric vector of length two containing the minimum value of
score and the maximum value of score. Default is \code{c(0, 1)}. The score is a
DisGeNet metric that estimates the strength of the association between
the variant and the disease}
}
\value{
A \code{data.frame} containing details about disease-SNPs and the associated
differentially expressed miRNAs.

The \code{variant} column contains SNP identifiers, the \code{disease} column
represents the disease name, the \code{mirnaGene} columns specifies the DE-miRNA
gene present, and the \code{allele} column shows the possible alleles of the
variant. Then, there are the \code{chromosome} and \code{position} columns, that give
the genomic coordinates of the SNP, and the \code{genesPresent} column that
indicates the genes present in the genomic trait in which the SNP is
located. Further, there are the \code{mirnaStart}, \code{mirnaEnd} and \code{mirnaStrand}
columns that provide genomic details of the differentially expressed miRNAs
affected by disease-SNPs. Finally, there are the \code{variantDSI}, \code{variantDPI},
\code{ei} and \code{score} columns which contain different association metrics
obtained from DisGeNet.
}
\description{
This function allows to identify genomic variants, positioned within
differentially expressed miRNA genes, that are associated with a
particular disease of interest. To do so, this function uses the DisGeNet
database to retrieve SNPs-disease associations, and then retains only
SNPs that affect DE-miRNA genes.
}
\details{
SNPs occurring within miRNAs may have important effects on the biological
function of these transcripts. Indeed, a SNP present within a miRNA might
alter the spectrum of targets regulated by a particular microRNA.

To retrieve disease-SNPs, this function uses DisGeNet, which is one of the
largest public databases containing information about human gene-disease
associations and human variants-disease associations. For variants, it
integrates information from multiple sources, such as GWAS catalogues and
biomedical literature. The \code{database} parameter specifies the database from
which the function obtains disease-associated SNPs. The options include:
\itemize{
\item \code{CLINVAR}, to use ClinVar;
\item \code{UNIPROT}, to use the Universal Protein Resource;
\item \code{GWASCAT}, to use the NHGRI-EBI GWAS Catalog;
\item \code{GWASDB}, to use GWAS Database GWASdb;
\item \code{BEFREE}, to use text mining data, generated using BeFree System;
\item \code{CURATED}, to use expert curated, human databases,
\item \code{ALL}, to use all the above mentioned databases together (default).
}

After running this function, the user can use the \code{\link[=mirVariantPlot]{mirVariantPlot()}}
function to produce a trackplot for visualizing the genomic presence of
SNPs within miRNA genes.
}
\note{
To retrieve disease-associated SNPs, this function makes use of the
\code{disgenet2r} package.
}
\examples{
# search disease
cuis <- searchDisease("Alzheimer's disease")
disId <- cuis[1, 1]

# retrieve associated SNPs
association <- findMirnaSNPs(obj, disId)
association

}
\references{
Piñero, J., Ramírez-Anguita, J. M., Saüch-Pitarch, J., Ronzano, F., Centeno,
E., Sanz, F., & Furlong, L. I. (2020). The DisGeNET knowledge platform for
disease genomics: 2019 update. Nucleic Acids Research, 48(D1), D845-D855.
}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
