% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{geneSet}
\alias{geneSet}
\title{Extract the gene-sets used for functional enrichment analyses}
\usage{
geneSet(object)
}
\arguments{
\item{object}{An object of class
\code{\link[=FunctionalEnrichment-class]{FunctionalEnrichment}} containing
enrichment results}
}
\value{
A \code{list} containing the gene-sets.
}
\description{
This function accesses the \code{geneSet} slot of a
\code{\link[=FunctionalEnrichment-class]{FunctionalEnrichment}} object and returns
a \code{list} with the collection of genes used for the enrichment.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# perform GSEA with KEGG
de_enr <- enrichGenes(obj, database = "KEGG")

# extract the gene-sets
gs <- geneSet(de_enr)

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
