% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{integrationDotplot}
\alias{integrationDotplot}
\title{Display integrated miRNA-mRNA augmented pathways in a dotplot}
\usage{
integrationDotplot(
  object,
  showTerms = 10,
  ordBy = "score",
  sizeBy = "ratio",
  colBy = "padj",
  title = NULL
)
}
\arguments{
\item{object}{An object of class
\code{\link[=IntegrativePathwayAnalysis-class]{IntegrativePathwayAnalysis}}}

\item{showTerms}{It is the number of top significant pathways to show or,
alternatively, a character vector indicating the pathways to plot.
Default is \code{10}}

\item{ordBy}{The parameter used to set the x-axis scale. It must be one of
\code{ratio}, \code{padj}, \code{pval} and \code{score} (default). See the \emph{details} section
for further information}

\item{sizeBy}{The parameter used to set the size scale. It must be one of
\code{ratio} (default), \code{padj}, \code{pval} and \code{score}. See the \emph{details} section
for further information}

\item{colBy}{The parameter used to set the color scale. It must be one of
\code{ratio}, \code{padj} (default), \code{pval} and \code{score}. See the \emph{details} section
for further information}

\item{title}{The title of the plot. Default is \code{NULL} not to include a plot
title}
}
\value{
A \code{ggplot} graph with a dotplot of integrated pathways.
}
\description{
This function produces a dotplot that depicts the results of a
topologically-aware integrative pathway analysis (TAIPA) carried out through
the \code{\link[=topologicalAnalysis]{topologicalAnalysis()}} function.
}
\details{
When producing the dotplot with default values, significant pathways
are ordered on the x-axis on the basis of their pathway score computed by
\code{\link[=topologicalAnalysis]{topologicalAnalysis()}}. The higher is this score, and the more affected a
pathway result between biological conditions. Moreover, the size of each
dot is equal to the ratio between the number of nodes for which a
measurement is available, and the total number of nodes. Finally, the color
scale of dots is relative to the adjusted p-values of each pathway. However,
the user can change the behavior of this function, by changing \code{ordBy},
\code{sizeBy}, and \code{colBy} parameters.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# perform integration analysis with default settings
obj <- mirnaIntegration(obj)

# perform the integrative pathway analysis with default settings
#ipa <- topologicalAnalysis(obj)

# access the results of pathway analysis
#integratedPathways(ipa)

# create a dotplot of integrated pathways
#integrationDotplot(ipa)

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
