% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IntegrativePathwayAnalysis-class}
\alias{IntegrativePathwayAnalysis-class}
\alias{integratedPathways,IntegrativePathwayAnalysis-method}
\alias{integrationDatabase,IntegrativePathwayAnalysis-method}
\alias{augmentedPathways,IntegrativePathwayAnalysis-method}
\title{The \code{IntegrativePathwayAnalysis} class}
\usage{
\S4method{integratedPathways}{IntegrativePathwayAnalysis}(object)

\S4method{integrationDatabase}{IntegrativePathwayAnalysis}(object)

\S4method{augmentedPathways}{IntegrativePathwayAnalysis}(object)
}
\description{
This class stores the output of integrative multi-omic pathway analyses.
In particular, the slots of this class are suitable to represent the results
of topologically-aware integrative pathway analyses (TAIPA) returned from
the \code{\link[=topologicalAnalysis]{topologicalAnalysis()}} function.
}
\details{
\subsection{Topologically-Aware Integrative Pathway Analysis (TAIPA)}{

This analysis aims to identify the biological pathways that result affected
by miRNA and mRNA dysregulations. In this analysis, biological pathways are
retrieved from a pathway database such as KEGG, and the interplay between
miRNAs and genes is then added to the networks. Each network is defined as
a graph \eqn{G(N,E)}, where \eqn{N} represents nodes, and \eqn{E} represents
the relationships between nodes. For each pathway, a score is a computed as:

\deqn{s = \frac{1}{N}\cdot \sum_{i = 0}^{N}\sum_{j = 0}^{U}\beta_{i\to j}
\left( \Delta E_i\cdot \Delta E_j\right)\,,}

where \eqn{N} is the total number of nodes in the pathway, \eqn{U} is the
total number of interactors for each node, \eqn{\Delta E} is the log2 fold
change of each node, and \eqn{\beta} is an interaction factor that
represents the directionality of a specific interaction (+1 for activation,
-1 for inhibition).

The higher is the score, and the more affected is a pathway. Then, to
compute the statistical significance of each pathway score, a permutation
procedure is applied. Indeed, for each pathway log2 fold changes are
permuted several times, and for each permutation pathway score is calculated.
In the end, the p-value is defined as the fraction of permutations that
reported a higher pathway score than the observed one.
}

\subsection{\code{data} slot}{

The \code{data} slot of this class consists in a \code{data.frame} object with seven
columns, namely:
\itemize{
\item \code{pathway}, which indicates the name of the biological network;
\item \code{considered.nodes}, which specifies the number of nodes with expression
measurement available;
\item \code{total.nodes}, which states the total number of nodes for this pathway;
\item \code{score}, which expresses the score of each individual pathway
\item \code{normalized.score}, the score of each pathway normalized by the total
number of nodes;
\item \code{P.Val}, the p-value of each pathway;
\item \code{Adj.P.Val}, the p-value adjusted for multiple testing.
}
}
}
\section{Functions}{
\itemize{
\item \code{integratedPathways(IntegrativePathwayAnalysis)}: Access the results of
integrative miRNA-mRNA pathway analysis

\item \code{integrationDatabase(IntegrativePathwayAnalysis)}: View the database used for
the integrative pathway analysis

\item \code{augmentedPathways(IntegrativePathwayAnalysis)}: Extract the list of biological
networks augmented with miRNA-mRNA interactions

}}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{data.frame} object that contains the results of the
integrative pathway analysis. See the \emph{details} section for further details}

\item{\code{method}}{The method used for the analysis}

\item{\code{organism}}{The name of the organism under consideration (e.g.
\verb{Homo sapiens})}

\item{\code{database}}{The name of the database used for retrieving biological
pathways (e.g. \code{KEGG})}

\item{\code{pCutoff}}{A \code{numeric} value defining the threshold used for
statistical significance (e.g. \code{0.05})}

\item{\code{pAdjustment}}{A \code{character} indicating the method used to correct
p-values for multiple testing (e.g. \code{fdr})}

\item{\code{pathways}}{A \code{list} of \code{graph} objects containing the biological
networks retrieved from \code{database}, and augmented with
miRNA-mRNA interactions}

\item{\code{minPc}}{The minimum percentage of measured features that a pathway must
have for being considered in the analysis}

\item{\code{nPerm}}{The number of permutation used for assessing the statistical
significance of each pathway}
}}

\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
