% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{listCategories}
\alias{listCategories}
\title{List the available miEAA 2.0 categories for an organism}
\usage{
listCategories(organism = "Homo sapiens", mirnaType = "mature")
}
\arguments{
\item{organism}{The name of the organism under consideration. It must be one
of: \verb{Homo sapiens}, \verb{Mus musculus}, \verb{Rattus norvegicus},
\verb{Arabidopsis thaliana}, \verb{Bos taurus}, \verb{Caenorhabditis elegans},
\verb{Drosophila melanogaster}, \verb{Danio rerio}, \verb{Gallus gallus}, \verb{Sus scrofa}}

\item{mirnaType}{The type of miRNA specie: use \code{mature} for the enrichment
of mature microRNAs; use \code{precursor}, for the enrichment analysis of miRNA
precursors}
}
\value{
A named vector containing the IDs of miEAA 2.0 categories. The user can
then decide the categories to use with \code{\link[=enrichMirnas]{enrichMirnas()}} and \code{\link[=gseaMirnas]{gseaMirnas()}}
functions.
}
\description{
This function retrieves a list of miEAA 2.0 categories that are available
for a specified organism. The name of these categories can be later used
with \code{\link[=enrichMirnas]{enrichMirnas()}} and \code{\link[=gseaMirnas]{gseaMirnas()}} functions to perform miRNA
enrichment analysis with the desired category.
}
\note{
This function uses the package \link{rbioapi} to interface miEAA 2.0 database
for retrieving the available categories for different organisms.
}
\examples{
# list all the available categories for Mus musculus
available_categories <- listCategories(organism = "Mus musculus")

}
\references{
Moosa Rezwani, Ali Akbar Pourfathollah, Farshid Noorbakhsh, rbioapi:
user-friendly R interface to biologic web services’ API, Bioinformatics,
Volume 38, Issue 10, 15 May 2022, Pages 2952–2953,
\link{https://doi.org/10.1093/bioinformatics/btac172}

Fabian Kern, Tobias Fehlmann, Jeffrey Solomon, Louisa Schwed, Nadja Grammes,
Christina Backes, Kendall Van Keuren-Jensen, David Wesley Craig, Eckart
Meese, Andreas Keller, miEAA 2.0: integrating multi-species microRNA
enrichment analysis and workflow management systems, Nucleic Acids Research,
Volume 48, Issue W1, 02 July 2020, Pages W521–W528,
\link{https://doi.org/10.1093/nar/gkaa309}
}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
