% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadExamples}
\alias{loadExamples}
\title{Create example \code{\link[=MirnaExperiment-class]{MirnaExperiment}} objects}
\usage{
loadExamples()
}
\value{
An example \code{MirnaExperiment} object.
}
\description{
This helper function allows to create a
\code{\link[=MirnaExperiment-class]{MirnaExperiment}} containing miRNA and gene
expression data deriving from Riesco-Eizaguirre et al (2015).
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
