% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/visualization.R
\name{mirnaDotplot}
\alias{mirnaDotplot}
\alias{mirnaDotplot,MirnaEnrichment-method}
\alias{mirnaDotplot,MirnaGsea-method}
\title{Create a dotplot for miRNA enrichment analysis}
\usage{
mirnaDotplot(
  object,
  showTerms = 10,
  splitDir = FALSE,
  ordBy = "fold",
  sizeBy = "Observed",
  colBy = "P.adjusted"
)

\S4method{mirnaDotplot}{MirnaEnrichment}(
  object,
  showTerms = 10,
  splitDir = FALSE,
  ordBy = "fold",
  sizeBy = "Observed",
  colBy = "P.adjusted"
)

\S4method{mirnaDotplot}{MirnaGsea}(
  object,
  showTerms = 10,
  splitDir = FALSE,
  ordBy = "fold",
  sizeBy = "Observed",
  colBy = "P.adjusted"
)
}
\arguments{
\item{object}{An object of class \code{\link{MirnaEnrichment}} or \code{\link{MirnaGsea}}
containing miRNA enrichment results}

\item{showTerms}{It is the number of top enriched terms to show or,
alternatively, a character vector indicating the enriched terms to plot.
Default is \code{10}}

\item{splitDir}{Logical, if \code{TRUE} the resulting plot will be divided in
two columns on the basis of enrichment direction (enrichment and depletion).
Default is \code{FALSE}}

\item{ordBy}{The parameter used to set the x-axis scale. It must be one of
\code{fold} (default), \code{P.adjusted}, \code{P.value} and \code{Observed}}

\item{sizeBy}{The parameter used to set the size scale. It must be one of
\code{fold}, \code{P.adjusted}, \code{P.value} and \code{Observed} (default)}

\item{colBy}{The parameter used to set the color scale. It must be one of
\code{fold}, \code{P.adjusted} (default), \code{P.value} and \code{Observed}}
}
\value{
A \code{ggplot} graph with a dotplot of enrichment results.
}
\description{
This function produces a dotplot to show the results of miRNA enrichment
analysis. This function can be used to plot results of both
over-representation analysis (ORA), carried out with \code{\link[=enrichMirnas]{enrichMirnas()}}, and
gene set enrichment analysis (GSEA), performed with \code{\link[=gseaMirnas]{gseaMirnas()}}.
}
\note{
For objects of class \code{\link{MirnaEnrichment}}, the \code{"fold"} value in \code{ordBy},
\code{sizeBy} and \code{colBy} refers to fold enrichment, which corresponds to the
ratio between observed and expected hits for a given category. This means
that the higher the fold enrichment is, the more unlikely it is that the
enrichment was reported by chance. Instead, for objects of class
\code{\link{MirnaGsea}} the \code{"fold"} value refers to the average absolute fold change
of observed miRNAs in a given category.
}
\examples{
# perform a GSEA with the miEAA category miRWalk
gse <- gseaMirnas(obj, organism = "Homo sapiens",
category = "miRWalk_GO_mature")

# extract results
res <- enrichmentResults(gse)

# plot results
mirnaDotplot(gse)

}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
