% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/enrichment.R, R/show-methods.R
\docType{class}
\name{MirnaEnrichment-class}
\alias{MirnaEnrichment-class}
\alias{[,MirnaEnrichment,character,missing,ANY-method}
\alias{show,MirnaEnrichment-method}
\title{The \code{MirnaEnrichment} class}
\usage{
\S4method{[}{MirnaEnrichment,character,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{show}{MirnaEnrichment}(object)
}
\description{
This function extends and adapts the \code{\link[DOSE:enrichResult-class]{DOSE::enrichResult}} class
in order to make it suitable for handling miRNA enrichment results.
}
\section{Methods (by generic)}{
\itemize{
\item \code{x[i}: Subset a \code{\link{MirnaEnrichment}} object that
contains multiple miEAA 2.0 categories into one that holds just the
specified category

\item \code{show(MirnaEnrichment)}: Show method for objects of class MirnaEnrichment

}}
\section{Slots}{

\describe{
\item{\code{result}}{A \code{data.frame} object holding the output of enrichment analysis}

\item{\code{pvalueCutoff}}{A \code{numeric} value defining the threshold used for
statistical significance in the enrichment analysis (e.g. \code{0.05})}

\item{\code{pAdjustMethod}}{A \code{character} indicating the method used to correct
p-values for multiple testing (e.g. \code{fdr})}

\item{\code{qvalueCutoff}}{A \code{numeric} value defining the q-value threshold used
in the enrichment analysis (e.g. \code{0.2})}

\item{\code{organism}}{The name of the organism under consideration (e.g.
\verb{Homo sapiens})}

\item{\code{ontology}}{The name of the miEAA 2.0 category used to perform the
enrichment analysis (e.g. \code{miRWalk_GO_mature})}

\item{\code{gene}}{A \code{character} vector containing the list of miRNAs used for the
enrichment}

\item{\code{keytype}}{The type of miRNA IDs used. For example \verb{miRBase v22}}

\item{\code{universe}}{The background universe of miRNAs used for the
over-representation analysis (ORA). Typically, this is equal to the complete
list of miRNAs assayed}

\item{\code{gene2Symbol}}{Mapping of genes to symbols, if needed}

\item{\code{geneSets}}{Gene sets}

\item{\code{readable}}{Logical flag of gene ID in symbol or not}

\item{\code{termism}}{Similarity between terms}

\item{\code{method}}{Method for calculating the similarity between nodes}

\item{\code{dr}}{Dimension reduction result}
}}

\references{
Guangchuang Yu, Li-Gen Wang, Guang-Rong Yan, Qing-Yu He. DOSE: an
R/Bioconductor package for Disease Ontology Semantic and Enrichment
analysis. Bioinformatics 2015 31(4):608-609
}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
