% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/enrichment.R, R/show-methods.R
\docType{class}
\name{MirnaGsea-class}
\alias{MirnaGsea-class}
\alias{[,MirnaGsea,character,missing,missing-method}
\alias{show,MirnaGsea-method}
\title{The \code{MirnaGsea} class}
\usage{
\S4method{[}{MirnaGsea,character,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{show}{MirnaGsea}(object)
}
\arguments{
\item{x}{An object of class \code{\link[=MirnaGsea-class]{MirnaGsea}} containing
enrichment results of multiple miEAA 2.0 categories}

\item{i}{A valid miEAA 2.0 category name}

\item{j}{Missing}

\item{...}{Additional arguments not used}

\item{drop}{Missing}

\item{object}{An object of class \code{\link[=MirnaGsea-class]{MirnaGsea}}}
}
\description{
This class extends and adapts the
\code{\link[DOSE:gseaResult-class]{gseaResult-class}} in order to make it
suitable for handling miRNA gene set enrichment analysis (GSEA) results.
}
\section{Functions}{
\itemize{
\item \code{x[i}: Subset a \code{\link[=MirnaGsea-class]{MirnaGsea}} object
that contains multiple miEAA 2.0 categories into one that holds just the
specified category

\item \code{show(MirnaGsea)}: Show method for objects of class MirnaGsea

}}
\section{Slots}{

\describe{
\item{\code{result}}{A \code{data.frame} object holding the output of enrichment analysis}

\item{\code{pvalueCutoff}}{A \code{numeric} value defining the threshold used for
statistical significance in the enrichment analysis (e.g. \code{0.05})}

\item{\code{pAdjustMethod}}{A \code{character} indicating the method used to correct
p-values for multiple testing (e.g. \code{fdr})}

\item{\code{organism}}{The name of the organism under consideration (e.g.
\verb{Homo sapiens})}

\item{\code{ontology}}{The name of the miEAA 2.0 category used to perform the
enrichment analysis (e.g. \code{miRWalk_GO_mature})}

\item{\code{gene}}{A \code{character} vector containing the list of miRNAs used for the
enrichment}

\item{\code{lfc}}{A \code{numeric} vector containing the list of log2 fold changes
relative to miRNAs listed in the \code{gene} slot}

\item{\code{keytype}}{The type of miRNA IDs used. For example \verb{miRBase v22}}

\item{\code{gene2Symbol}}{Mapping of genes to symbols, if needed}

\item{\code{geneSets}}{Gene sets}

\item{\code{setType}}{Set type}

\item{\code{geneList}}{Order rank gene list}

\item{\code{permScores}}{Permutation scores}

\item{\code{params}}{Parameters}

\item{\code{readable}}{Logical flag of gene ID in symbol or not}

\item{\code{dr}}{Dimension reduction result}
}}

\references{
Guangchuang Yu, Li-Gen Wang, Guang-Rong Yan, Qing-Yu He. DOSE: an
R/Bioconductor package for Disease Ontology Semantic and Enrichment
analysis. Bioinformatics 2015 31(4):608-609
}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
