% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{mirnaPathway}
\alias{mirnaPathway}
\title{Explore the relationships between miRNAs and genes within biological pathways}
\usage{
mirnaPathway(mirnaObj, pathway, database = "KEGG", organism = "Homo sapiens")
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{pathway}{The name of a biological pathway to integrate with miRNA
expression. It must be a valid pathway name contained in KEGG, Reactome or
WikiPathways. For example: \code{"Phosphorylation of CD3 and TCR zeta chains"}
contained in Reactome database}

\item{database}{The name of the database used for retrieving the biological.
pathway. It must be one of: \code{KEGG}, \code{Reactome}, and \code{WikiPathways}.
Default is \code{KEGG}}

\item{organism}{The name of the organism under consideration. The different
databases have different supported organisms. To see the list of supported
organisms for a given database, use the \code{\link[=supportedOrganisms]{supportedOrganisms()}} function.
Default is \verb{Homo sapiens}}
}
\value{
A \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}} object containing the biological pathway
integrated with gene and miRNA expression. To plot this network object, you
can use the \code{\link[=visualizeNetwork]{visualizeNetwork()}} function.
}
\description{
This function is useful to integrate biological pathways with microRNAs. In
particular, this function retrieves a user-defined pathway from one of the
available databases (KEGG, Reactome and WikiPathways), and integrate every
gene of the pathway with the miRNAs that regulate those members. The
resulting network can then be used with the \link{visualizeNetwork} function
to visualize the relationships between miRNA and gene expression.
}
\note{
This function uses the \code{graphite} package to retrieve biological pathways as
\code{graph} objects from KEGG, Reactome and WikiPathways.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# perform the integration analysis between miRNAs and their targets
obj <- mirnaIntegration(obj)

# explore Thyroid hormone synthesis pathway in KEGG
net <- mirnaPathway(obj,
pathway = "Thyroid hormone synthesis", organism = "Homo sapiens",
database = "KEGG")

# visualize miRNA-gene network
visualizeNetwork(net)

}
\references{
Sales, G., Calura, E., Cavalieri, D. et al. graphite - a Bioconductor
package to convert pathway topology to gene network.
BMC Bioinformatics 13, 20 (2012).
\url{https://doi.org/10.1186/1471-2105-13-20}

Minoru Kanehisa, Susumu Goto, KEGG: Kyoto Encyclopedia of Genes and Genomes,
Nucleic Acids Research, Volume 28, Issue 1, 1 January 2000, Pages 27–30,
\url{https://doi.org/10.1093/nar/28.1.27}

G. Joshi-Tope, M. Gillespie, I. Vastrik, P. D'Eustachio, E. Schmidt, B.
de Bono, B. Jassal, G.R. Gopinath, G.R. Wu, L. Matthews, S. Lewis, E.
Birney, L. Stein, Reactome: a knowledgebase of biological pathways, Nucleic
Acids Research, Volume 33, Issue suppl_1, 1 January 2005, Pages D428–D432,
\url{https://doi.org/10.1093/nar/gki072}

Marvin Martens, Ammar Ammar, Anders Riutta, Andra Waagmeester, Denise N
Slenter, Kristina Hanspers, Ryan A. Miller, Daniela Digles, Elisson N Lopes,
Friederike Ehrhart, Lauren J Dupuis, Laurent A Winckers, Susan L Coort, Egon
L Willighagen, Chris T Evelo, Alexander R Pico, Martina Kutmon,
WikiPathways: connecting communities, Nucleic Acids Research, Volume 49,
Issue D1, 8 January 2021, Pages D613–D621,
\url{https://doi.org/10.1093/nar/gkaa1024}
}
\seealso{
For additional details on how to handle \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}} objects,
refer to the \code{tidygraph} package.
}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
