% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{mirnaRidgeplot}
\alias{mirnaRidgeplot}
\title{Create a ridgeplot to display the results of miRNA GSEA analysis}
\usage{
mirnaRidgeplot(mirnaGsea, showTerms = 10, colBy = "P.adjusted", title = NULL)
}
\arguments{
\item{mirnaGsea}{A \code{\link[=MirnaGsea-class]{MirnaGsea}} object containing the
results of gene set enrichment analysis obtained with the \code{\link[=gseaMirnas]{gseaMirnas()}}
function}

\item{showTerms}{It is the number of top enriched terms to show or,
alternatively, a character vector indicating the enriched terms to plot.
Default is \code{10}}

\item{colBy}{The parameter used to set the color scale. It must be one of
\code{P.adjusted} (default), \code{P.value} and \code{Observed}}

\item{title}{The title of the plot. Default is \code{NULL} not to include a plot
title}
}
\value{
An object of class \code{ggplot} containing the ridgeplot of miRNA GSEA results.
}
\description{
This function creates a ridgeplot that is useful for showing the results
of a miRNA GSEA analysis. The output of this function is a plot where
enrichmed terms found with \code{\link[=gseaMirnas]{gseaMirnas()}} function are visualized on the
basis of miRNA fold changes. In particular, the plot displays enriched
terms on the y-axis and miRNA fold changes on the x-axis. The resulting
area represents the density of fold changes belonging to miRNAs annotated
to that category.
}
\examples{
# perform miRNA GSEA analysis
gse_res <- gseaMirnas(obj, organism = "Homo sapiens",
category = "miRWalk_GO_mature")

# plot results as a ridgeplot
mirnaRidgeplot(gse_res)

}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
