% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllClasses.R
\name{mirnaTargetsIntegration}
\alias{mirnaTargetsIntegration}
\alias{mirnaTargetsIntegration,MirnaExperiment-method}
\title{Explore the results of the integration analysis between miRNAs and genes}
\usage{
mirnaTargetsIntegration(object)

\S4method{mirnaTargetsIntegration}{MirnaExperiment}(object)
}
\arguments{
\item{object}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object containing
miRNA and gene data}
}
\value{
A \code{data.frame} object containing the results of the integration analysis.
}
\description{
After performing the integration analysis between miRNA and gene expression
values with the \code{\link[=integrateMirnaTargets]{integrateMirnaTargets()}} function, the results are stored
in the \code{mirnaTargetsIntegration} slot of a
\code{\link[=MirnaExperiment-class]{MirnaExperiment}} object and can be explored with
this function.
}
\examples{
# perform Kendall's correlation analysis with tau > 0.8 and p < 0.05
obj <- integrateMirnaTargets(obj, test = "correlation",
corMethod = "kendall", corCutoff = 0.8)

# visualize the results of correlation analysis
res <- mirnaTargetsIntegration(obj)
res

}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
