% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{searchDisease}
\alias{searchDisease}
\title{Search for disease CUI identifiers}
\usage{
searchDisease(diseaseName)
}
\arguments{
\item{diseaseName}{The name of a particular disease
(ex. \verb{Creutzfeldt-Jakob disease}).}
}
\value{
A \code{data.frame} object containing CUI identifiers in the first column, and
full disease names in the second column.
}
\description{
This function allows to retrieve the UMLS Concept Unique Identifier (CUI)
of a particular disease. This identifier is then needed to use the function
\code{\link[=findMirnaSNPs]{findMirnaSNPs()}}.
}
\note{
To retrieve CUIs for specific diseases, this function makes use of the
\code{disgenet2r} package.
}
\examples{
# search the CUI identifier of Creutzfeldt-Jakob disease
cui <- searchDisease("Creutzfeldt-Jakob disease")
cui

# search the CUI identifier of Alzheimer's disease
cui <- searchDisease("Alzheimer's disease")
cui

}
\references{
Piñero, J., Ramírez-Anguita, J. M., Saüch-Pitarch, J., Ronzano, F., Centeno,
E., Sanz, F., & Furlong, L. I. (2020). The DisGeNET knowledge platform for
disease genomics: 2019 update. Nucleic Acids Research, 48(D1), D845-D855.
}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
