% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{setTargets}
\alias{setTargets}
\title{Use custom miRNA-target interactions}
\usage{
setTargets(mirnaObj, tg)
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{tg}{A \code{data.frame} object specifying miRNA-target interactions. Check
the \strong{details section} for the format}
}
\value{
A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object containing miRNA targets
stored in the \code{targets} slot. Results can be accessed with the
\code{\link[=mirnaTargets]{mirnaTargets()}} function.
}
\description{
Apart from the \code{\link[=getTargets]{getTargets()}} function, users can use their own custom
list of miRNA-target interactions. This function allows to directly add
custom interactions to an existing
\code{\link[=MirnaExperiment-class]{MirnaExperiment}} object.
}
\details{
This function requires interactions to be passed as a \code{data.frame} object
with two column, one for miRNAs and one for genes. The column names of this
\code{data.frame} must include \code{MicroRNA} and \code{Gene.Symbol}.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# define targets
tg <- data.frame(MicroRNA = c("hsa-miR-1179", "hsa-miR-1179"),
                 Gene.Symbol = c("ACVR2A", "ABI2"))

# set targets
obj <- setTargets(obj, tg)

# print targets
mirnaTargets(obj)

}
\references{
Ronchi, J., & Foti, M. (2026). MIRit: An integrative R framework for the
identification of impaired miRNA–mRNA regulatory networks in complex
diseases. Bioinformatics Advances, vbag042.
\url{https://doi.org/10.1093/bioadv/vbag042}
}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
