% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllClasses.R
\name{significantGenes}
\alias{significantGenes}
\alias{significantGenes,MirnaExperiment-method}
\title{Get the IDs of statistically differentially expressed genes}
\usage{
significantGenes(object)

\S4method{significantGenes}{MirnaExperiment}(object)
}
\arguments{
\item{object}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object containing
miRNA and gene data}
}
\value{
A \code{character} vector of gene symbols (e.g. 'TP53', 'FOXP2', 'TIGAR'
'CASP1'...).
}
\description{
This function is an accessor of the \code{significantGenes} slot of a
\code{\link[=MirnaExperiment-class]{MirnaExperiment}} object and can be used to obtain
the IDs of statistically differentially expressed genes.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# extract significant DEGs
sigGenes <- significantGenes(obj)

}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
