% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllClasses.R
\name{significantMirnas}
\alias{significantMirnas}
\alias{significantMirnas,MirnaExperiment-method}
\title{Get the IDs of statistically differentially expressed miRNAs}
\usage{
significantMirnas(object)

\S4method{significantMirnas}{MirnaExperiment}(object)
}
\arguments{
\item{object}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object containing
miRNA and gene data}
}
\value{
A \code{character} vector of miRNA IDs (e.g. 'hsa-miR-16-5p',
'hsa-miR-29a-3p'...).
}
\description{
This function is an accessor of the \code{significantMirnas} slot of a
\code{\link[=MirnaExperiment-class]{MirnaExperiment}} object and can be used to obtain
the IDs of statistically differentially expressed miRNAs.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# extract significant DE-miRNAs
sigMirnas <- significantMirnas(obj)

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
