% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological-integration.R
\name{topologicalAnalysis}
\alias{topologicalAnalysis}
\title{Perform a topologically-aware integrative pathway analysis (TAIPA)}
\usage{
topologicalAnalysis(
  mirnaObj,
  database = "KEGG",
  organism = "Homo sapiens",
  pCutoff = 0.05,
  pAdjustment = "fdr",
  nPerm = 10000,
  minPc = 10,
  BPPARAM = bpparam()
)
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{database}{The name of the database to use. It must be one of: \code{KEGG},
\code{Reactome}, and \code{WikiPathways}. Default is \code{KEGG}}

\item{organism}{The name of the organism under consideration. The different
databases have different supported organisms. To see the list of supported
organisms for a given database, use the \code{\link[=supportedOrganisms]{supportedOrganisms()}} function.
Default specie is \verb{Homo sapiens}}

\item{pCutoff}{The adjusted p-value cutoff to use for statistical
significance. The default value is \code{0.05}}

\item{pAdjustment}{The p-value correction method for multiple testing. It
must be one of: \code{fdr} (default), \code{BH}, \code{none}, \code{holm}, \code{hochberg}, \code{hommel},
\code{bonferroni}, \code{BY}}

\item{nPerm}{The number of permutation used for assessing the statistical
significance of each pathway. Default is 10000. See the \emph{details} section
for additional information}

\item{minPc}{The minimum percentage of measured features that a pathway must
have for being considered in the analysis. Default is 10. See the \emph{details}
section for additional information}

\item{BPPARAM}{The desired parallel computing behavior. This paramete
defaults to \code{BiocParallel::bpparam()}, but this can be edited. See
\code{\link[BiocParallel:register]{BiocParallel::bpparam()}} for information on parallel computing in R}
}
\value{
An object of class
\code{\link[=IntegrativePathwayAnalysis-class]{IntegrativePathwayAnalysis}} that stores
the results of the analysis. See the relative help page for further details.
}
\description{
This function allows to perform an integrative pathway analysis that aims
to identify the biological networks that are most affected by miRNomic
and transcriptomic dysregulations. Briefly, influential miRNA-mRNA
interactions, identified by the \code{\link[=mirnaIntegration]{mirnaIntegration()}} function, are added to
biological pathways retrieved from a pathway database such as \code{KEGG},
\code{WikiPathways} and \code{Reactome}. Then, a score that estimates the degree of
impairment is calculated for each pathway, and statistical significance is
calculated through a permutation test. The main advantages of this method
is that it doesn't require matched samples, and that it allows to perform
an integrative miRNA-mRNA pathway analysis that take into account the
topology of biological pathways and miRNA-mRNA interactions. See the
\emph{details} section for additional information.
}
\details{
\subsection{Topologically-Aware Integrative Pathway Analysis (TAIPA)}{

This analysis aims to identify the biological pathways that result affected
by miRNA and mRNA dysregulations. In this analysis, biological pathways are
retrieved from a pathway database such as KEGG, and the interplay between
miRNAs and genes is then added to the networks. Each network is defined as
a graph \eqn{G(N,E)}, where \eqn{N} represents nodes, and \eqn{E} represents
the relationships between nodes. For each pathway, a score is a computed as:

\deqn{s = \frac{1}{N}\cdot \sum_{i = 0}^{N}\sum_{j = 0}^{U}\beta_{i\to j}
\left( \Delta E_i\cdot \Delta E_j\right)\,,}

where \eqn{N} is the total number of nodes in the pathway, \eqn{U} is the
total number of interactors for each node, \eqn{\Delta E} is the log2 fold
change of each node, and \eqn{\beta} is an interaction factor that
represents the direction of a specific interaction (+1 for activation,
-1 for inhibition).

The higher is the score, and the more affected is a pathway. Then, to
compute the statistical significance of each pathway score, a permutation
procedure is applied. Indeed, for each pathway log2 fold changes are
permuted several times, and for each permutation pathway score is calculated.
In the end, the p-value is defined as the fraction of permutations that
reported a higher pathway score than the observed one.
}

\subsection{Implementation details}{

To create augmented pathways, this function uses the \code{graphite} R package
to download biological networks from the above mentioned database. Then,
each pathway is converted to a graph object, and significant miRNA-mRNA
interactions are added to the network. Further, edge weights are added
according to interaction type.

At this point, biological pathways with few nodes measured are excluded
from this analysis. This is required because, during differential expression
analysis, lowly expressed features are removed. Therefore, some pathways
might result significantly affected even if only 1\% of nodes is perturbed.
The default behavior is to exclude pathways with less than 10\% of
representation (\code{minPc = 10}).

After this normalization step, the score of each pathway is calculated.
For computational efficiency, pathway score computation has been implemented
in C++ language.

Moreover, to define the statistical significance of each network, a
permutation test is applied following the number of permutations specified
with \code{nPerm}. The default setting is to perform 10000 permutations. The
higher is the number of permutations, the more stable are the calculated
p-values, even though the time needed will increase. In this regard, since
computing pathway score for 10000 networks for each pathway is
computationally intensive, parallel computing has been employed to reduce
running time. The user can modify the parallel computing behavior by
specifying the \code{BPPARAM} parameter. See \code{\link[BiocParallel:register]{BiocParallel::bpparam()}} for
further details.
}
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# perform integration analysis with default settings
obj <- mirnaIntegration(obj)

# perform the integrative pathway analysis with default settings
#ipa <- topologicalAnalysis(obj)

# access the results of pathway analysis
#integratedPathways(ipa)

# create a dotplot of integrated pathways
#integrationDotplot(ipa)

# explore a specific biological network
#visualizeNetwork(ipa, "Thyroid hormone synthesis")

}
\references{
Sales, G., Calura, E., Cavalieri, D. et al. graphite - a Bioconductor
package to convert pathway topology to gene network.
BMC Bioinformatics 13, 20 (2012),
\url{https://doi.org/10.1186/1471-2105-13-20}.
}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
