% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichGenes}
\alias{enrichGenes}
\title{Perform an over-representation analysis of differentially expressed genes}
\usage{
enrichGenes(
  mirnaObj,
  database = "GO",
  organism = "Homo sapiens",
  simplifyGO = TRUE,
  pCutoff = 0.05,
  qCutoff = 0.2,
  pAdjustment = "fdr",
  ...
)
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{database}{The name of the database used for the enrichment analysis.
It must be one of: \code{GO}, \code{KEGG}, \code{Reactome}, \code{DisGeNet} and \code{WikiPathways}.
Default is \code{GO}}

\item{organism}{The name of the organism under consideration. The different
databases have different supported organisms. To see the list of supported
organisms for a given database, use the \code{\link[=supportedOrganisms]{supportedOrganisms()}} function.
Default is \verb{Homo sapiens}}

\item{simplifyGO}{Logical, whether to apply the
\code{\link[clusterProfiler:simplify-methods]{clusterProfiler::simplify()}} function in \link{clusterProfiler} package to
reduce the redundancy of GO terms in the results (default is \code{TRUE}).
This argument is only considered when \code{database} is set to \code{GO}. For further
information see the \emph{details} section}

\item{pCutoff}{The adjusted p-value cutoff to use for statistical
significance. The default value is \code{0.05}}

\item{qCutoff}{The q-value cutoff to use for statistical significance
The default value is \code{0.2}}

\item{pAdjustment}{The p-value correction method for multiple testing. It
must be one of: \code{fdr} (default), \code{BH}, \code{none}, \code{holm}, \code{hochberg}, \code{hommel},
\code{bonferroni}, \code{BY}}

\item{...}{Other parameters that can be given to the enrichment functions}
}
\value{
A \code{list} object with two elements, namely 'upregulated' and 'downregulated',
containing enrichment results of upregulated and downregulated genes,
respectively. Each element of this \code{list} consists of an object of class
\link[=`DOSE:enrichResult-class`]{enrichResult} containing the outcome
of the enrichment analysis of DE-genes. To access the enrichment results
as a standard \code{data.frame}, you can use the \code{as.data.frame()} function.
The user can also refer to the documentation of \link{clusterProfiler} for
analyzing and plotting results:
\link{http://yulab-smu.top/biomedical-knowledge-mining-book/index.html}.
}
\description{
This function allows to enrich upregulated and downregulated genes through
an over-representation analysis (ORA). This can be done to investigate the
biological effects of gene expression dysregulations. The enrichment
analysis can be carried out using different databases, namely Gene Ontology
(GO) - Biological Process, Kyoto Encyclopedia of Genes and Genomes (KEGG),
Reactome, DisGeNet and WikiPathways.
}
\details{
If the enrichment analysis is performed using \code{GO} database,
the \code{\link[clusterProfiler:simplify-methods]{clusterProfiler::simplify()}} method contained in the \link{clusterProfiler}
package may be used. This is particularly useful since the structure of GO
database is highly redundant. Therefore, to remove some of the redundancy
from the resulting enriched GO terms, the parameter \code{simplifyGO} can be set
to \code{TRUE}.
}
\note{
The over-representation analysis is implemented in this function through
the packages \link{clusterProfiler}, \link{ReactomePA} and \link{DOSE}.
}
\examples{
# perform enrichment analysis of DE genes with Reactome
de_enr <- enrichGenes(obj, database = "Reactome")

# extract the enrichment of upregulated genes
enr_up <- de_enr[["upregulated"]]

# extract enrichment results as a data.frame
enr_df <- as.data.frame(enr_up)
enr_df

# create a dotplot of enriched terms
enrichplot::dotplot(enr_up)

}
\references{
T Wu, E Hu, S Xu, M Chen, P Guo, Z Dai, T Feng, L Zhou, W Tang, L Zhan, X
Fu, S Liu, X Bo, and G Yu. clusterProfiler 4.0: A universal enrichment tool
for interpreting omics data. The Innovation. 2021, 2(3):100141

Guangchuang Yu, Qing-Yu He. ReactomePA: an R/Bioconductor package for
reactome pathway analysis and visualization. Molecular BioSystems 2016,
12(2):477-479

Guangchuang Yu, Li-Gen Wang, Guang-Rong Yan, Qing-Yu He. DOSE: an
R/Bioconductor package for Disease Ontology Semantic and Enrichment
analysis. Bioinformatics 2015 31(4):608-609

Ashburner, M., Ball, C., Blake, J. et al. Gene Ontology: tool for the
unification of biology. Nat Genet 25, 25–29 (2000).
\link{https://doi.org/10.1038/75556}

Minoru Kanehisa, Susumu Goto, KEGG: Kyoto Encyclopedia of Genes and Genomes,
Nucleic Acids Research, Volume 28, Issue 1, 1 January 2000, Pages 27–30,
\link{https://doi.org/10.1093/nar/28.1.27}

G. Joshi-Tope, M. Gillespie, I. Vastrik, P. D'Eustachio, E. Schmidt, B.
de Bono, B. Jassal, G.R. Gopinath, G.R. Wu, L. Matthews, S. Lewis, E.
Birney, L. Stein, Reactome: a knowledgebase of biological pathways, Nucleic
Acids Research, Volume 33, Issue suppl_1, 1 January 2005, Pages D428–D432,
\link{https://doi.org/10.1093/nar/gki072}

Piñero, J., Ramírez-Anguita, J. M., Saüch-Pitarch, J., Ronzano, F., Centeno,
E., Sanz, F., & Furlong, L. I. (2020). The DisGeNET knowledge platform for
disease genomics: 2019 update. Nucleic Acids Research, 48(D1), D845-D855.

Marvin Martens, Ammar Ammar, Anders Riutta, Andra Waagmeester, Denise N
Slenter, Kristina Hanspers, Ryan A. Miller, Daniela Digles, Elisson N Lopes,
Friederike Ehrhart, Lauren J Dupuis, Laurent A Winckers, Susan L Coort, Egon
L Willighagen, Chris T Evelo, Alexander R Pico, Martina Kutmon,
WikiPathways: connecting communities, Nucleic Acids Research, Volume 49,
Issue D1, 8 January 2021, Pages D613–D621,
\link{https://doi.org/10.1093/nar/gkaa1024}
}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
