% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllClasses.R
\name{enrichmentResults}
\alias{enrichmentResults}
\alias{enrichmentResults,FunctionalEnrichment-method}
\title{Extract results from enrichment objects}
\usage{
enrichmentResults(object)

\S4method{enrichmentResults}{FunctionalEnrichment}(object)
}
\arguments{
\item{object}{An object of class
\code{\link[=FunctionalEnrichment-class]{FunctionalEnrichment}} containing
enrichment results.}
}
\value{
A \code{data.frame} object with the full results of the enrichment analysis.
}
\description{
This function allows to access the \code{data} slot of a
\code{\link[=FunctionalEnrichment-class]{FunctionalEnrichment}} object. This is useful
to take a closer look at all the enriched terms of an enrichment analysis. Nfor objects of classes
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# perform GSEA with KEGG database
gse <- enrichMirnas(obj, organism = "Homo sapiens",
database = "KEGG", method = "GSEA")

# extract results
res <- enrichmentResults(gse)

# plot results
enrichmentDotplot(gse)

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
