% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllClasses.R
\name{enrichmentResults}
\alias{enrichmentResults}
\alias{enrichmentResults,enrichResult-method}
\alias{enrichmentResults,MirnaEnrichment-method}
\alias{enrichmentResults,MirnaGsea-method}
\title{Extract results from enrichment objects}
\usage{
enrichmentResults(object)

\S4method{enrichmentResults}{enrichResult}(object)

\S4method{enrichmentResults}{MirnaEnrichment}(object)

\S4method{enrichmentResults}{MirnaGsea}(object)
}
\arguments{
\item{object}{An object of class \code{\link{MirnaEnrichment}} or \code{\link{MirnaGsea}}
containing miRNA enrichment results. Alternatively, object of class
\code{\link[DOSE:enrichResult-class]{DOSE::enrichResult}} or \code{\link[DOSE:gseaResult-class]{DOSE::gseaResult}} are also accepted}
}
\value{
A \code{data.frame} object with the full results of the enrichment analysis.
}
\description{
This function allows to access the \code{result} slot of \code{\link{MirnaEnrichment}} and
\code{\link{MirnaGsea}} objects. This is useful to take a closer look at all the
enriched terms of an enrichment analysis. Note that this function can also
extract the enrichment results for objects of classes
\code{\link[DOSE:enrichResult-class]{DOSE::enrichResult}} and \code{\link[DOSE:gseaResult-class]{DOSE::gseaResult}}.
}
\examples{
# perform a GSEA with the miEAA category miRWalk
gse <- gseaMirnas(obj, organism = "Homo sapiens",
category = "miRWalk_GO_mature")

# extract results
res <- enrichmentResults(gse)

# plot results
mirnaDotplot(gse)

}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
