% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllClasses.R
\name{geneDE}
\alias{geneDE}
\alias{geneDE,MirnaExperiment-method}
\title{Extract differentially expressed genes from a
\code{\link[=MirnaExperiment-class]{MirnaExperiment}} object}
\usage{
geneDE(object, onlySignificant = TRUE)

\S4method{geneDE}{MirnaExperiment}(object, onlySignificant = TRUE)
}
\arguments{
\item{object}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object containing
miRNA and gene data}

\item{onlySignificant}{Logical, if \code{TRUE} differential expression results
will be returned just for statistically significant genes, if \code{FALSE} the
full table of gene differential expression will be provided. Default is
\code{TRUE} to only report significant genes}
}
\value{
A \code{data.frame} with gene differential expression.
}
\description{
This function is an accessor for the \code{geneDE} slot of
\code{\link[=MirnaExperiment-class]{MirnaExperiment}}  class, and thus, it can be used
to retrieve the results of gene differential expression stored in a
\code{\link[=MirnaExperiment-class]{MirnaExperiment}} object.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# access gene differential expression of a MirnaExperiment object
sig <- geneDE(obj)
all <- geneDE(obj, onlySignificant = FALSE)

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
