% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{getTargets}
\alias{getTargets}
\title{Get microRNA targets}
\usage{
getTargets(
  mirnaObj,
  organism = "Homo sapiens",
  onlyValidated = FALSE,
  minPredicted = 3,
  minValidated = 1,
  topCutoff = 20
)
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{organism}{The name of the organism under consideration. It must be one
of: \verb{Homo sapiens}, \verb{Mus musculus}, \verb{Rattus norvegicus}}

\item{onlyValidated}{Logical, if \code{TRUE}, the function will retrieve only
experimentally validated targets; if \code{FALSE}, a combination of validated
and predicted targets will be used (default is \code{FALSE})}

\item{minPredicted}{Considers as true targets only those predicted by at
least n prediction algorithms. It should be within 1-8 (default is \code{3})}

\item{minValidated}{Considers as true targets only those present in more
than n databases with validated interactions. It should be within 1-3
(default is \code{1})}

\item{topCutoff}{This value corresponds to the percentage of top results
retrieved by the different prediction databases present in \code{multiMiR}.
Default is \code{20}}
}
\value{
A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object containing targets of
differentially expressed miRNAs. Targets are stored in the \code{targets} slot
and can be accessed with the \code{\link[=mirnaTargets]{mirnaTargets()}} function.
}
\description{
This function allows to obtain the targets of differentially expressed
miRNAs using the \code{multiMiR} package, which outputs microRNA targets from
several databases, including 8 prediction databases and 3 databases with
experimentally validated targets. The microRNA-mRNA interactions reported
by this function can derive by both predicted and validated interactions.
The user can decide whether to retrieve only validated targets or a
combination of predicted and validated targets. Moreover, for predicted
interactions, the user can decide either to consider as targets all
predicted targets or to include only those targets predicted by a minimum
of n prediction databases.
}
\examples{
# retrieve validated targets and targets predicted by at lest 3 databases
mirnaObj <- getTargets(mirnaObj = mirnaObj, onlyValidated = FALSE,
minPredicted = 3, minValidated = 1, topCutoff = 20)

}
\references{
Yuanbin Ru, Katerina J. Kechris, Boris Tabakoff, Paula Hoffman, Richard A.
Radcliffe, Russell Bowler, Spencer Mahaffey, Simona Rossi, George A. Calin,
Lynne Bemis, Dan Theodorescu; The multiMiR R package and database:
integration of microRNA–target interactions along with their disease and
drug associations. Nucleic Acids Res 2014; 42 (17): e133. doi:
\url{10.1093/nar/gku631}
}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
