% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrateMirnaTargets}
\alias{integrateMirnaTargets}
\title{Integrate microRNA and gene expression}
\usage{
integrateMirnaTargets(
  mirnaObj,
  test = "auto",
  pCutoff = 0.05,
  pAdjustment = "fdr",
  onlySignificant = TRUE,
  corMethod = "spearman",
  corCutoff = 0.5
)
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{test}{The statistical test to evaluate the association between miRNAs
and genes. It must be one of \code{auto} (default), to automatically determine
the appropriate statistical test; \code{correlation}, to perform a correlation
analysis; \code{fisher}, to perform a one-sided Fisher's exact test}

\item{pCutoff}{The adjusted p-value cutoff to use for statistical
significance. The default value is \code{0.05}}

\item{pAdjustment}{The p-value correction method for multiple testing. It
must be one of: \code{fdr} (default), \code{BH}, \code{none}, \code{holm}, \code{hochberg}, \code{hommel},
\code{bonferroni}, \code{BY}}

\item{onlySignificant}{Logical, whether to report only statistically
significant associations (default is \code{TRUE}). This parameter is considered
only if a one-sided Fisher's exact test is applied}

\item{corMethod}{The correlation method to be used for correlation analysis.
It must be one of: \code{spearman} (default), \code{pearson}, \code{kendall}. See the
'details' section for further information}

\item{corCutoff}{The minimum (negative) value of correlation coefficient to
consider meaningful a miRNA-target relationship. Default is \code{0.5}}
}
\value{
A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object containing integration
results. To access these results, the user can make use of the
\code{\link[=mirnaTargetsIntegration]{mirnaTargetsIntegration()}} function, which returns a \code{data.frame} object
with the details of miRNA-targets relationships. This \code{data.frame} differs
on the basis of the integration strategy used. For the one-sided Fisher's
exact test integration, it has seven columns:
\itemize{
\item \code{microRNA}: the miRNA ID;
\item \code{direction}: the fold change direction of the DE-miRNA (\code{upregulated} or
\code{downregulated});
\item \code{n_DE_targets}: represents the number of differentially expressed targets;
\item \code{n_NON_DE_targets}: represents the number of non differentially expressed
targets;
\item \code{Fisher.P.Val}: indicates the p-value resulting from the one-sided
Fisher's exact test;
\item \code{Fisher.Adjusted.P.Val}: contains the Fisher's exact test p-values
corrected for multiple testing;
\item \code{DE_targets}: contains the list of differentially expressed targtes whose
expression is negatively associated with miRNA expression.
Instead, when a correlation analysis is performed, \code{mirnaTargetsIntegration}
has seven columns:
\item \code{microRNA}: the miRNA ID;
\item \code{Target}: the correlated target gene;
\item \code{microRNA.Direction}: the fold change direction of the DE-miRNA;
\item \code{Pearson/Spearman/Kendall.Coeff}: the value of the correlation coefficient
used;
\item \code{Corr.P.Value}: the p-value resulting from the correlation analysis;
\item \code{Corr.Adjusted.P.Val}: contains the correlation p-values corrected for
multiple testing.
}
}
\description{
This function allows to test which microRNAs are significantly
associated/correlated with their targets. The principle is that, since the
biological role of miRNAs is mainly to negatively regulate gene expression
post-transcriptionally, the expression of a microRNA should be negatively
correlated with the expression of its targets. To test this assumption, this
function implements a correlation analysis, if miRNA and gene expression
data derive from the same samples (paired data), whereas, with unpaired data,
it uses a one-sided Fisher's exact to estimate if targets of down-regulated
miRNAs are enriched in up-regulated genes and vice versa. See the \emph{details}
section for further information.
}
\details{
As already pointed out, if miRNA and gene expression data derive from the
same samples, a correlation analysis is used. For evaluating these
relationships, the default method used is Spearman's correlation
coefficient, as:
\itemize{
\item it does not need normally distributed data;
\item it does not assume linearity;
\item it is much more resistant to outliers.
}

However, the user can also decide to use other correlation methods,
such as Pearson's and Kendall's correlation. Nevertheless, for NGS data
it may happen that a certain number of ties is present in the expression
values. This can be handled by \code{spearman} method as it computes a
tie-corrected version of Spearman's coefficients. However, another
correlation method that is suitable to perform rank correlation on tied data
is the Kendall's tau-b method, usable with \code{kendall}.

Regarding correlation direction, since miRNAs mainly act as negative
regulators, only negatively correlated miRNA-target pairs are evaluated, and
statistical significance is calculated through a one-tailed t-test.

Finally, if gene expression data and miRNA expression data derive from
different samples (unpaired data), a correlation analysis can't be
performed. However, a one-sided Fisher's exact test can be applied in these
cases to evaluate if targets of down-regulated miRNAs are statistically
enriched in up-regulated genes, and, conversely, if targets of up-regulated
miRNAs are statistically enriched in down-regulated genes.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# perform integration analysis with default settings
obj <- integrateMirnaTargets(obj)

# use the Fisher's exact test  with FDR < 0.05 as significance threshold
obj <- integrateMirnaTargets(obj, test = "fisher", pAdjustment = "fdr")

# perform Kendall's correlation analysis with tau > 0.8 and p < 0.05
obj <- integrateMirnaTargets(obj, test = "correlation",
corMethod = "kendall", corCutoff = 0.8)

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
