% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllClasses.R
\name{mirnaDE}
\alias{mirnaDE}
\alias{mirnaDE,MirnaExperiment-method}
\title{Extract differentially expressed miRNAs from a
\code{\link[=MirnaExperiment-class]{MirnaExperiment}} object}
\usage{
mirnaDE(object, onlySignificant = TRUE)

\S4method{mirnaDE}{MirnaExperiment}(object, onlySignificant = TRUE)
}
\arguments{
\item{object}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object containing
miRNA and gene data}

\item{onlySignificant}{Logical, if \code{TRUE} differential expression results
will be returned just for statistically significant miRNAs, if \code{FALSE} the
full table of miRNA differential expression will be provided. Default is
\code{TRUE} to only report significant miRNAs}
}
\value{
A \code{data.frame} with miRNA differential expression.
}
\description{
This function is an accessor for the \code{mirnaDE} slot of
\code{\link[=MirnaExperiment-class]{MirnaExperiment}} class, and thus, it can be used
to retrieve the results of miRNA differential expression stored in a
\code{\link[=MirnaExperiment-class]{MirnaExperiment}} object.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# access miRNA differential expression of a MirnaExperiment object
sig <- mirnaDE(obj)
all <- mirnaDE(obj, onlySignificant = FALSE)

}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
