% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/show-methods.R
\docType{class}
\name{MirnaExperiment-class}
\alias{MirnaExperiment-class}
\alias{show,MirnaExperiment-method}
\title{The 'MirnaExperiment' class}
\usage{
\S4method{show}{MirnaExperiment}(object)
}
\description{
This class extends the \code{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment}}
from the homonym package to provide the flexibility of handling genomic data
of both microRNAs and their targets, allowing to store information about
microRNA and gene expression, differential expression results, microRNA
targets and miRNA-gene integration analysis.

This class can be used to manage genomic data deriving from different
sources, like RNA-Seq, microarrays and mass spectrometry. Moreover, microRNA
and gene expression levels may originate from the same individuals
(paired samples) or from different subjects (unpaired samples).
}
\section{Functions}{
\itemize{
\item \code{show(MirnaExperiment)}: Show method for objects of class
MirnaExperiment

}}
\section{Slots}{

\describe{
\item{\code{ExperimentList}}{A \code{\link[MultiAssayExperiment:ExperimentList]{MultiAssayExperiment::ExperimentList}} class object
for each assay dataset}

\item{\code{colData}}{A \code{DataFrame} of all clinical/specimen data available across
experiments}

\item{\code{sampleMap}}{A \code{DataFrame} of translatable identifiers of samples and
participants}

\item{\code{metadata}}{Additional data describing the object}

\item{\code{drops}}{A metadata \code{list} of dropped information}

\item{\code{mirnaDE}}{A \code{data.frame} object containing miRNA differential expression
results}

\item{\code{geneDE}}{A \code{data.frame} object containing gene differential expression
results}

\item{\code{significantMirnas}}{A \code{character} vector containing the names of
significantly differentially expressed miRNAs}

\item{\code{significantGenes}}{A \code{character} vector containing the names of
significantly differentially expressed genes}

\item{\code{pairedSamples}}{A \code{logical} parameter that specifies whether miRNA and
gene expression measurements derive from the same individuals (\code{TRUE}) or
from different subjects (\code{FALSE})}

\item{\code{targets}}{A \code{data.frame} object containing miRNA-target pairs. This
slot is commonly populated by the \code{\link[=getTargets]{getTargets()}} function}

\item{\code{mirnaTargetsIntegration}}{A \code{data.frame} object containing the results
of the integration analysis between miRNA and gene expression values. This
slot is commonly populated by the \code{\link[=integrateMirnaTargets]{integrateMirnaTargets()}} function}
}}

\note{
To create a \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object, you can use
the \code{\link[=MirnaExperiment]{MirnaExperiment()}} constructor function, which allows to easily build
and verify a valid object starting from the results of differential
expression analysis.
}
\section{mirnaDE and geneDE}{


\code{mirnaDE} consists of a \code{data.frame} with five columns:
\itemize{
\item \code{ID}: indicates the name of the miRNA;
\item \code{logFC}: indicates the fold change of each miRNA in logarithmic scale;
\item \code{AveExpr}: represents the average expression of each miRNA;
\item \code{P.Value}: indicates the p-value resulting from differential expression
analysis;
\item \code{FDR}: contains the p-values adjusted for multiple testing.
}

\code{geneDE} consists of a \code{data.frame} with five columns:
\itemize{
\item \code{ID}: indicates the name of the gene;
\item \code{logFC}: indicates the fold change of each gene in logarithmic scale;
\item \code{AveExpr}: represents the average expression of each gene;
\item \code{P.Value}: indicates the p-value resulting from differential expression
analysis;
\item \code{FDR}: contains the p-values adjusted for multiple testing.
}
}

\section{significantMirnas and significantGenes}{


\code{significantMirnas} and \code{significantGenes} are \code{character} vectors of
miRNA/gene IDs that are considered as statistically differentially expressed.
The IDs contained in these vectors must be present in the \code{ID} columns of
\code{mirnaDE} and \code{geneDE}.
}

\section{pairedSamples}{


As already mentioned, \code{pairedSamples} must be \code{TRUE} when miRNA and gene
expression derive from the same subjects, while it must \code{FALSE} if this is
not the case.
}

\section{targets}{


\code{targets} is a \code{data.frame} with just two columns:
\itemize{
\item \code{mature_mirna_id}, which contains miRNA names; and
\item \code{target_symbol}, which indicates the target gene for the corresponding
miRNA.
}
}

\section{mirnaTargetsIntegration}{


Lastly, \code{mirnaTargetsIntegration} slot contains a \code{data.frame} that differs
on the basis of the integration strategy used. For the one-sided Fisher's
exact test integration, this \code{data.frame} has seven columns:
\itemize{
\item \code{microRNA}: the miRNA ID;
\item \code{direction}: the fold change direction of the DE-miRNA (\code{upregulated} or
\code{downregulated});
\item \code{n_DE_targets}: represents the number of differentially expressed targets;
\item \code{n_NON_DE_targets}: represents the number of non differentially expressed
targets;
\item \code{Fisher.P.Val}: indicates the p-value resulting from the one-sided
Fisher's exact test;
\item \code{Fisher.Adjusted.P.Val}: contains the Fisher's exact test p-values
corrected for multiple testing;
\item \code{DE_targets}: contains the list of differentially expressed targets whose
expression is negatively associated with miRNA expression.
Instead, when a correlation analysis is performed, \code{mirnaTargetsIntegration}
has seven columns:
\item \code{microRNA}: the miRNA ID;
\item \code{target}: the correlated target gene;
\item \code{microRNA.Direction}: the fold change direction of the DE-miRNA;
\item \code{Correlation.Type}: specifies the correlation direction, either \code{positive}
or \code{negative};
\item \code{Corr.Coefficient}: the value of the correlation coefficient used;
\item \code{Corr.P.Value}: the p-value resulting from the correlation analysis;
\item \code{Corr.Adjusted.P.Val}: contains the correlation p-values corrected for
multiple testing.
}
}

\section{ExperimentList}{


The \code{\link[MultiAssayExperiment]{ExperimentList}} slot is designed to contain results from
each experiment/assay. It contains a \linkS4class{SimpleList}.

}

\section{colData}{


The \code{colData} slot is a collection of primary specimen data valid
across all experiments. This slot is strictly of class
\code{\linkS4class{DataFrame}} but arguments for the constructor function
allow arguments to be of class \code{data.frame} and subsequently coerced.

}

\section{sampleMap}{


The \code{\link[MultiAssayExperiment]{sampleMap}} contains a \code{DataFrame} of translatable
identifiers of samples and participants or biological units. The standard
column names of the \code{sampleMap} are "assay", "primary", and "colname".
Note that the "assay" column is a factor corresponding to the names of each
experiment in the \code{ExperimentList}. In the case where these names do
not match between the \code{sampleMap} and the experiments, the documented
experiments in the \code{sampleMap} take precedence and experiments are
dropped by the harmonization procedure. The constructor function will
generate a \code{sampleMap} in the case where it is not provided and this
method may be a 'safer' alternative for creating the \code{MultiAssayExperiment}
(so long as the rownames are identical in the \code{colData}, if provided).
An empty \code{sampleMap} may produce empty experiments if the levels of the
"assay" factor in the \code{sampleMap} do not match the names in the
\code{ExperimentList}.

}

\references{
Marcel Ramos et al. Software For The Integration Of Multiomics Experiments
In Bioconductor. Cancer Research, 2017 November 1; 77(21); e39-42. DOI:
\link{10.1158/0008-5472.CAN-17-0344}
}
\seealso{
See \code{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment-class}}
for additional information.
}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
