% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotDE}
\alias{plotDE}
\title{Represent differentially expressed miRNAs/genes as boxplots, barplots or
violinplots}
\usage{
plotDE(
  mirnaObj,
  features,
  condition,
  graph = "boxplot",
  showSignificance = TRUE,
  starSig = TRUE,
  pCol = "adj.P.Val",
  sigOffset = 0.9,
  sigLabelSize = 7,
  digits = 3,
  colorScale = NULL
)
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{features}{A character vector containing the genes/miRNAs to plot}

\item{condition}{It must be the column name of a variable specified in the
metadata (colData) of a \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object;
or, alternatively, it must be a character/factor object that specifies
group memberships (eg. c("healthy, "healthy", "disease", "disease"))}

\item{graph}{The type of plot to produce. It must be one of \code{boxplot}
(default), \code{barplot}, \code{violinplot}}

\item{showSignificance}{Logical, whether to display statistical significance
or not. Default is TRUE}

\item{starSig}{Logical, whether to represent statistical significance through
stars. Default is TRUE, and the significance scale is: \* for $p < 0.05$, \**
for $p < 0.01$, \*** for $p < 0.001$, and \**** for $p < 0.0001$. If
\code{starSig} is set to FALSE, p-values or adjusted p-values will be reported on
the plot as numbers}

\item{pCol}{The statistics used to evaluate comparison significance. It must
be one of \code{P.Value}, to use unadjusted p-values, and \code{adj.P.Val} (default),
to use p-values corrected for multiple testing}

\item{sigOffset}{The distance between the different brackets used to
show statistical significance. Default is 0.9, but the user can increment it
to enlarge the distance between significance brackets}

\item{sigLabelSize}{The size for the labels used to show statistical
significance. Default is 7, which is well suited for representing p-values
as significance stars. However, if \code{starSig} is set to FALSE, the user might
have to downsize this parameter}

\item{digits}{The number of digits to show when p-values are reported as
numbers (when \code{starSig} is FALSE). Default is 3}

\item{colorScale}{It must be a named character vector where values
correspond to R colors, while names coincide with the genes/miRNAs specified
in the \code{condition} parameter (eg. c("hsa-miR-34a-5p" = "blue",
"PAX8" = "red")). Default is NULL, in order to use the default color scale}
}
\value{
An object of class \code{ggplot} containing the plot.
}
\description{
This function is able to produce boxplots, barplots and violinplots that are
useful to visualize miRNA and gene differential expression. The user just
has to provide a vector of interesting miRNA/genes that he wants to plot
(e.g. "hsa-miR-34a-5p", "hsa-miR-146b-5p", "PAX8"). The chart type can be
specified through the \code{graph} parameter.
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# produce a boxplot for PAX8 and miR-34a-5p
plotDE(obj, features = c("hsa-miR-34a-5p", "PAX8"), condition = "disease")

# produce a barplot for PAX8 and miR-34a-5p without significance
plotDE(obj, features = c("hsa-miR-34a-5p", "PAX8"), condition = "disease",
graph = "barplot", showSignificance = FALSE)

# produce a violinplot for BCL2
plotDE(obj, features = "BCL2", condition = "disease", graph = "violinplot")

}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
