% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{supportedOrganisms}
\alias{supportedOrganisms}
\title{Get the list of supported organisms for a given database}
\usage{
supportedOrganisms(database)
}
\arguments{
\item{database}{The name of the database used for the enrichment analysis.
It must be one of: \code{GO}, \code{KEGG}, \code{Reactome}, \code{DisGeNet} and \code{WikiPathways}}
}
\value{
A \code{character} listing all the supported organisms for the database
specified by the user.
}
\description{
This function provides the list of supported organisms for different
databases, namely Gene Ontology (GO), Kyoto Encyclopedia of Genes and
Genomes (KEGG), Reactome, DisGeNet and WikiPathways.
}
\examples{
# get the supported organisms for GO database
supportedOrganisms("GO")

# get the supported organisms for Reactome
supportedOrganisms("Reactome")

}
\references{
Ashburner, M., Ball, C., Blake, J. et al. Gene Ontology: tool for the
unification of biology. Nat Genet 25, 25–29 (2000).
\link{https://doi.org/10.1038/75556}

Minoru Kanehisa, Susumu Goto, KEGG: Kyoto Encyclopedia of Genes and Genomes,
Nucleic Acids Research, Volume 28, Issue 1, 1 January 2000, Pages 27–30,
\link{https://doi.org/10.1093/nar/28.1.27}

G. Joshi-Tope, M. Gillespie, I. Vastrik, P. D'Eustachio, E. Schmidt, B.
de Bono, B. Jassal, G.R. Gopinath, G.R. Wu, L. Matthews, S. Lewis, E.
Birney, L. Stein, Reactome: a knowledgebase of biological pathways, Nucleic
Acids Research, Volume 33, Issue suppl_1, 1 January 2005, Pages D428–D432,
\link{https://doi.org/10.1093/nar/gki072}

Piñero, J., Ramírez-Anguita, J. M., Saüch-Pitarch, J., Ronzano, F., Centeno,
E., Sanz, F., & Furlong, L. I. (2020). The DisGeNET knowledge platform for
disease genomics: 2019 update. Nucleic Acids Research, 48(D1), D845-D855.

Marvin Martens, Ammar Ammar, Anders Riutta, Andra Waagmeester, Denise N
Slenter, Kristina Hanspers, Ryan A. Miller, Daniela Digles, Elisson N Lopes,
Friederike Ehrhart, Lauren J Dupuis, Laurent A Winckers, Susan L Coort, Egon
L Willighagen, Chris T Evelo, Alexander R Pico, Martina Kutmon,
WikiPathways: connecting communities, Nucleic Acids Research, Volume 49,
Issue D1, 8 January 2021, Pages D613–D621,
\link{https://doi.org/10.1093/nar/gkaa1024}
}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
