% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological-integration.R
\name{topologicalAnalysis}
\alias{topologicalAnalysis}
\title{Perform a topologically-aware integrative pathway analysis (TAIPA)}
\usage{
topologicalAnalysis(
  mirnaObj,
  database = "KEGG",
  organism = "Homo sapiens",
  pCutoff = 0.05,
  pAdjustment = "max-T",
  nPerm = 10000,
  minPc = 10,
  BPPARAM = bpparam()
)
}
\arguments{
\item{mirnaObj}{A \code{\link[=MirnaExperiment-class]{MirnaExperiment}} object
containing miRNA and gene data}

\item{database}{The name of the database to use. It must be one of: \code{KEGG},
\code{Reactome}, and \code{WikiPathways}. Default is \code{KEGG}}

\item{organism}{The name of the organism under consideration. The different
databases have different supported organisms. To see the list of supported
organisms for a given database, use the \code{\link[=supportedOrganisms]{supportedOrganisms()}} function.
Default specie is \verb{Homo sapiens}}

\item{pCutoff}{The adjusted p-value cutoff to use for statistical
significance. The default value is \code{0.05}}

\item{pAdjustment}{The p-value correction method for multiple testing. It
must be one of: \code{max-T} (default), \code{fdr}, \code{BH}, \code{none}, \code{holm}, \code{hochberg},
\code{hommel}, \code{bonferroni}, \code{BY}}

\item{nPerm}{The number of permutation used for assessing the statistical
significance of each pathway. Default is 10000. See the \emph{details} section
for additional information}

\item{minPc}{The minimum percentage of measured features that a pathway must
have for being considered in the analysis. Default is 10. See the \emph{details}
section for additional information}

\item{BPPARAM}{The desired parallel computing behavior. This paramete
defaults to \code{BiocParallel::bpparam()}, but this can be edited. See
\code{\link[BiocParallel:register]{BiocParallel::bpparam()}} for information on parallel computing in R}
}
\value{
An object of class
\code{\link[=IntegrativePathwayAnalysis-class]{IntegrativePathwayAnalysis}} that stores
the results of the analysis. See the relative help page for further details.
}
\description{
This function allows to perform an integrative pathway analysis that aims
to identify the biological networks that are most affected by miRNomic
and transcriptomic dysregulations. Briefly, influential miRNA-mRNA
interactions, identified by the \code{\link[=mirnaIntegration]{mirnaIntegration()}} function, are added to
biological pathways retrieved from a pathway database such as \code{KEGG},
\code{WikiPathways} and \code{Reactome}. Then, a score that estimates the degree of
impairment is calculated for each pathway, and statistical significance is
calculated through a permutation test. The main advantages of this method
are that it doesn't require matched samples, and that it allows to perform
an integrative miRNA-mRNA pathway analysis that takes into account the
topology of biological networks. See the \emph{details} section for
additional information.
}
\details{
\subsection{Topologically-Aware Integrative Pathway Analysis (TAIPA)}{

This analysis aims to identify the biological pathways that result affected
by miRNA and mRNA dysregulations. In this analysis, biological pathways are
retrieved from a pathway database such as KEGG, and the interplay between
miRNAs and genes is then added to the networks. Each network is defined as
a graph \eqn{G(V, E)}, where \eqn{V} represents nodes, and \eqn{E}
represents the relationships between nodes.

Then, nodes that are not significantly differentially expressed are assigned
a weight \eqn{w_i = 1}, whereas differentially expressed nodes are assigned
a weight \eqn{w_i = \left| \Delta E_i \right|}, where \eqn{\Delta E_i} is
the linear fold change of the node. Moreover, to consider the biological
interaction between two nodes, namely \eqn{i} and \eqn{j}, we define an
interaction parameter \eqn{\beta_{i \rightarrow j} = 1} for activation
interactions and \eqn{\beta_{i \rightarrow j} = -1} for repression
interactions. Subsequently, the concordance coefficient
\eqn{\gamma_{i \rightarrow j}} is defined as:

\deqn{\gamma_{i \rightarrow j} = \begin{cases} \beta_{i \rightarrow j}
&\text{if } sign(\Delta E_i) = sign(\Delta E_j) \\ - \beta_{i \rightarrow j}
&\text{if } sign(\Delta E_i) \not= sign(\Delta E_j) \end{cases}\,.}

Later in the process, a breadth-first search (BFS) algorithm is applied to
topologically sort pathway nodes so that each individual node occurs after
all its upstream nodes. Nodes within cycles are considered leaf nodes. At
this point, a node score \eqn{\phi} is calculated for each pathway node
\eqn{i} as:

\deqn{\phi_i = w_i + \sum_{j=1}^{U} \gamma_{i \rightarrow j} \cdot k_j\,.}

where \eqn{U} represents the number of upstream nodes,
\eqn{\gamma_{i \rightarrow j}} denotes the concordance coefficient, and
\eqn{k_j} is a propagation factor defined as:

\deqn{k_j = \begin{cases} w_j &\text{if } \phi_j = 0 \\ \phi_j &\text{if }
\phi_j \not = 0 \end{cases}\,.}

Finally, the pathway score \eqn{\Psi} is calculated as:

\deqn{\Psi = \frac{1 - M}{N} \cdot \sum_{i=1}^{N} \phi_i\,,}

where \eqn{M} represents the proportion of miRNAs in the pathway, and
\eqn{N} represents the total number of nodes in the network.

Then, to compute the statistical significance of each pathway score, a
permutation procedure is applied. Later, both observed pathway scores and
permuted scores are standardized by subtracting the mean score of the
permuted sets \eqn{\mu_{\Psi_P}} and then dividing by the standard deviation
of the permuted scores \eqn{\sigma_{\Psi_P}}.

Finally, the p-value is defined as the fraction of permutations that
reported a higher normalized pathway score than the observed one. In the
end, p-values are corrected for multiple testing either through the max-T
procedure (default option) which is particularly suited for permutation
tests, or through the standard multiple testing approaches.
}

\subsection{Implementation details}{

To create augmented pathways, this function uses the \code{graphite} R package
to download biological networks from the above mentioned database. Then,
each pathway is converted to a \code{graph} object, and significant miRNA-mRNA
interactions are added to the network. Further, edge weights are added
according to interaction type.

At this point, biological pathways with few nodes measured are excluded
from this analysis. This is required because, during differential expression
analysis, lowly expressed features are removed. Therefore, some pathways
might result significantly affected even if only 1\% of nodes is perturbed.
The default behavior is to exclude pathways with less than 10\% of
representation (\code{minPc = 10}).

After this normalization step, the score of each pathway is calculated.
For computational efficiency, pathway score computation has been implemented
in C++ language.

Moreover, to define the statistical significance of each network, a
permutation test is applied following the number of permutations specified
with \code{nPerm}. The default setting is to perform 10000 permutations. The
higher is the number of permutations, the more stable are the calculated
p-values, even though the time needed will increase. In this regard, since
computing pathway score for 10000 networks for each pathway is
computationally intensive, parallel computing has been employed to reduce
running time. The user can modify the parallel computing behavior by
specifying the \code{BPPARAM} parameter. See \code{\link[BiocParallel:register]{BiocParallel::bpparam()}} for
further details.
}
}
\examples{
# load example MirnaExperiment object
obj <- loadExamples()

# perform integration analysis with default settings
obj <- mirnaIntegration(obj)

# perform the integrative pathway analysis with default settings
#ipa <- topologicalAnalysis(obj)

# access the results of pathway analysis
#integratedPathways(ipa)

# create a dotplot of integrated pathways
#integrationDotplot(ipa)

# explore a specific biological network
#visualizeNetwork(ipa, "Thyroid hormone synthesis")

}
\references{
Sales, G., Calura, E., Cavalieri, D. et al. graphite - a Bioconductor
package to convert pathway topology to gene network.
BMC Bioinformatics 13, 20 (2012),
\url{https://doi.org/10.1186/1471-2105-13-20}.

Peter H. Westfall and S. Stanley Young. Resampling-Based Multiple Testing:
Examples and Methods for p-Value Adjustment. John Wiley & Sons.
ISBN 978-0-471-55761-6.
}
\author{
Jacopo Ronchi, \email{jacopo.ronchi@unimib.it}
}
