% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualizeNetwork}
\alias{visualizeNetwork}
\title{Visualize the relationships between miRNAs and genes in a biological pathway}
\usage{
visualizeNetwork(
  pathGraph,
  onlyIntegrated = FALSE,
  layout = "stress",
  node = "box",
  size = 3,
  title = NULL,
  edgesCol = "gray",
  bindingEdgesCol = "gray",
  stress.text.distance = 0.2,
  circle.text.distance = 0.1
)
}
\arguments{
\item{pathGraph}{A \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}} object created with the
\code{\link[=mirnaPathway]{mirnaPathway()}} function.}

\item{onlyIntegrated}{Logical. It must be set to \code{TRUE} for building the
network only for integrated targets, i.e. those targets whose expression is
statistically associated/correlated with miRNA expression. Instead, to build
the complete biological pathway, \code{onlyIntegrated} must be set to
\code{FALSE} (default)}

\item{layout}{It must be either \code{stress} (default), to build the network
through an optimized Kamada-Kawai algorithm, or \code{circle}, to represent the
pathway through a circular layout}

\item{node}{It must be either \code{box} (default), to represent nodes as
boxed labels, or \code{points}, to display nodes as points with the gene/miRNA
identifier}

\item{size}{The size of node elements in the graph. Default is \code{3}}

\item{title}{The title of the plot (e.g. "Thyroid hormone signaling
pathway)". Default is \code{NULL} not to include a plot title}

\item{edgesCol}{It must be an R color name that specifies the color of
interaction arrows in the network. Default is \code{gray}. All available colors
can be listed with \code{\link[=colors]{colors()}}}

\item{bindingEdgesCol}{It must be an R color name that specifies the color
of chemical binding arrows in the network. Default is \code{gray}. All available
colors can be listed with \code{\link[=colors]{colors()}}}

\item{stress.text.distance}{The distance between activation/inhibition
arrow caps and nodes for the \code{stress} design. Default is \code{0.2}}

\item{circle.text.distance}{The distance between activation/inhibition
arrow caps and nodes for the \code{circle} design. Default is \code{0.1}}
}
\value{
A \code{ggraph} object representing the network graph. For instructions on how
to handle and edit this object, please refer to the \link{ggraph} package.
}
\description{
This function is used to plot the network object created with the
\code{\link[=mirnaPathway]{mirnaPathway()}} function. The resulting network represents the interactions
occurring within a biological pathway alongside with regulating microRNAs
that affect pathway members. Moreover, this function visually displays the
network with gene and miRNA expression levels, so that the user can easily
understand the biological implications of miRNA dysregulations.
}
\details{
The \code{layout} parameter determines the layout of the integrated network. If
this parameter is set to \code{stress}, this function adopts an optimized version
of the stress-minimization algorithm (known as Kamada-Kawai algorithm).
Otherwise, if \code{circle} layout is used, biological pathways will be
displayed as a circular network. Both of these layout algorithms are
implemented in the \link{ggraph} package.

The \code{stress} layout is suitable for small pathways, but it can get messed
for larger biological networks. In these cases, the \code{circle} layout
performs better.
}
\examples{
# explore a TCR-associated pathway present in Reactome
net <- mirnaPathway(obj,
pathway = "Phosphorylation of CD3 and TCR zeta chains", organism = "Homo
sapiens", database = "Reactome", onlyIntegrated = TRUE)

# visualize miRNA-gene network
visualizeNetwork(net)

# visualize network with a circular layout and with points
visualizeNetwork(net, layout = "circle", node = "point")

}
\author{
Jacopo Ronchi, \email{j.ronchi2@campus.unimib.it}
}
